+++ srcs/ft_putendl.c
@@ -3,18 +3,18 @@
 /*                                                        :::      ::::::::   */
 /*   ft_putendl.c                                       :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: ekruhliu <marvin@42.fr>                    +#+  +:+       +#+        */
+/*   By: ekruhliu <ekruhliu@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2017/11/09 17:57:55 by ekruhliu          #+#    #+#             */
-/*   Updated: 2017/11/09 17:57:56 by ekruhliu         ###   ########.fr       */
+/*   Updated: 2017/11/15 12:37:52 by ekruhliu         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-void	ft_putendl(char const *s)
+void		ft_putendl(char const *s)
 {
-	int i;
+	size_t	i;
 
 	i = 0;
 	if (s != NULL)
