+++ srcs/ft_putendl_fd.c
@@ -3,18 +3,18 @@
 /*                                                        :::      ::::::::   */
 /*   ft_putendl_fd.c                                    :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: ekruhliu <marvin@42.fr>                    +#+  +:+       +#+        */
+/*   By: ekruhliu <ekruhliu@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2017/11/09 19:32:41 by ekruhliu          #+#    #+#             */
-/*   Updated: 2017/11/09 19:32:42 by ekruhliu         ###   ########.fr       */
+/*   Updated: 2017/11/15 12:38:05 by ekruhliu         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-void	ft_putendl_fd(char const *s, int fd)
+void		ft_putendl_fd(char const *s, int fd)
 {
-	int i;
+	size_t	i;
 
 	i = 0;
 	if (s != NULL)
