+++ srcs/ft_putstr_fd.c
@@ -3,18 +3,18 @@
 /*                                                        :::      ::::::::   */
 /*   ft_putstr_fd.c                                     :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: ekruhliu <marvin@42.fr>                    +#+  +:+       +#+        */
+/*   By: ekruhliu <ekruhliu@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2017/11/09 19:22:47 by ekruhliu          #+#    #+#             */
-/*   Updated: 2017/11/09 19:22:48 by ekruhliu         ###   ########.fr       */
+/*   Updated: 2017/11/15 12:38:33 by ekruhliu         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-void	ft_putstr_fd(char const *s, int fd)
+void		ft_putstr_fd(char const *s, int fd)
 {
-	int i;
+	size_t	i;
 
 	i = 0;
 	if (s != NULL)
