+++ srcs/ft_strchr.c
@@ -3,19 +3,19 @@
 /*                                                        :::      ::::::::   */
 /*   ft_strchr.c                                        :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: ekruhliu <marvin@42.fr>                    +#+  +:+       +#+        */
+/*   By: ekruhliu <ekruhliu@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2017/10/30 11:54:15 by ekruhliu          #+#    #+#             */
-/*   Updated: 2017/10/30 11:54:16 by ekruhliu         ###   ########.fr       */
+/*   Updated: 2017/11/16 12:49:10 by ekruhliu         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-char		*ft_strchr(const char *str, int c)
+char			*ft_strchr(const char *str, int c)
 {
-	int		i;
-	char	*s;
+	size_t		i;
+	char		*s;
 
 	i = 0;
 	s = (char*)str;
