+++ srcs/ft_strequ.c
@@ -3,18 +3,18 @@
 /*                                                        :::      ::::::::   */
 /*   ft_strequ.c                                        :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: ekruhliu <marvin@42.fr>                    +#+  +:+       +#+        */
+/*   By: ekruhliu <ekruhliu@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2017/11/06 13:00:04 by ekruhliu          #+#    #+#             */
-/*   Updated: 2017/11/06 13:00:06 by ekruhliu         ###   ########.fr       */
+/*   Updated: 2017/11/15 12:44:39 by ekruhliu         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-int	ft_strequ(char const *s1, char const *s2)
+int			ft_strequ(char const *s1, char const *s2)
 {
-	int i;
+	size_t	i;
 
 	i = 0;
 	if (s1 != NULL && s2 != NULL)
