+++ srcs/ft_strnequ.c
@@ -3,18 +3,18 @@
 /*                                                        :::      ::::::::   */
 /*   ft_strnequ.c                                       :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: ekruhliu <marvin@42.fr>                    +#+  +:+       +#+        */
+/*   By: ekruhliu <ekruhliu@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2017/11/06 13:58:35 by ekruhliu          #+#    #+#             */
-/*   Updated: 2017/11/06 13:58:36 by ekruhliu         ###   ########.fr       */
+/*   Updated: 2017/11/15 12:45:29 by ekruhliu         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-int	ft_strnequ(char const *s1, char const *s2, size_t n)
+int			ft_strnequ(char const *s1, char const *s2, size_t n)
 {
-	int i;
+	size_t	i;
 
 	i = 0;
 	if (s1 != NULL && s2 != NULL)
